+++ CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to the project will be documented in this file.
 
 #### API breaking changes
 
-N/A
+- Xcode 10 and Swift 4.2 support
 
 #### Enhancements